refactor: make history an alias for sessions
